Write the app driven by the guide. Then correct the guide appropriately.
Using the guide, I drafted the application to perform as expected. There were small corrections, but
it basically worked. Then, I copy-and-pasted the code fragments back into the guide to sync them up.
Based on adjustments to the app,

In the process, I uncovered a ConcurrentModificationException that happens when you iterate over the
map of users and try to delete an entry. I thought I had mitigated the risk by using an iterator, but
that tip appears to be false. Flagged it with a TODO in the code as well as the guide, so that this issue
doesn't get lost.